changed the structure of directories and edited comments
