--- WAFL-DETR/filter/__init__.py
@@ -1,1 +1,0 @@
-from . import generate_noniid
