+++ WAFL-DETR/src/slurm-1325644.out
@@ -1,0 +1,18 @@
+/home/u01071/anaconda3/lib/python3.11/site-packages/torchvision/models/_utils.py:208: UserWarning: The parameter 'pretrained' is deprecated since 0.13 and may be removed in the future, please use 'weights' instead.
+  warnings.warn(
+/home/u01071/anaconda3/lib/python3.11/site-packages/torchvision/models/_utils.py:223: UserWarning: Arguments other than a weight enum or `None` for 'weights' are deprecated since 0.13 and may be removed in the future. The current behavior is equivalent to passing `weights=ResNet50_Weights.IMAGENET1K_V1`. You can also use `weights=ResNet50_Weights.DEFAULT` to get the most up-to-date weights.
+  warnings.warn(msg)
+Namespace(lr=1e-05, lr_backbone=1e-06, batch_size=2, weight_decay=0.0001, epochs=200, lr_drop=1000, clip_max_norm=0.1, frozen_weights=None, tle=None, he=None, backbone='resnet50', dilation=False, position_embedding='sine', enc_layers=6, dec_layers=6, dim_feedforward=2048, hidden_dim=256, dropout=0.1, nheads=8, num_queries=100, pre_norm=False, masks=False, aux_loss=True, set_cost_class=1, set_cost_bbox=5, set_cost_giou=2, mask_loss_coef=1, dice_loss_coef=1, bbox_loss_coef=5, giou_loss_coef=2, eos_coef=0.1, num_classes=10, dataset_file='custom', coco_path='./data/custom/', coco_panoptic_path=None, remove_difficult=False, output_dir='outputs', device='cuda', seed=42, resume='detr-r50_no-class-head.pth', start_epoch=0, eval=False, num_workers=2, world_size=1, dist_url='env://', num_clients=10, noniid_ratio=90, iid_setting=False, no_exchange=False, preself_epochs=100, topology='line', resume_from_preselftrained='', resume_from_wafl='', distributed=None)
+number of params: [41281551, 41281551, 41281551, 41281551, 41281551, 41281551, 41281551, 41281551, 41281551, 41281551]
+Traceback (most recent call last):
+  File "/home/u01071/wafl/WAFL-DETR/src/main_wafl.py", line 399, in <module>
+    main(args)
+  File "/home/u01071/wafl/WAFL-DETR/src/main_wafl.py", line 198, in main
+    dataset_train = build_dataset(image_set='train', args=args)
+                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+  File "/home/u01071/wafl/WAFL-DETR/src/datasets/__init__.py", line 27, in build_dataset
+    return build_custom(image_set, args)
+           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+  File "/home/u01071/wafl/WAFL-DETR/src/datasets/custom.py", line 12, in build
+    assert root.exists(), f'provided path {root} to custom dataset does not exist'
+AssertionError: provided path data/custom to custom dataset does not exist
