+++ WAFL-DETR/src/util/box_visualize.py
@@ -7,7 +7,6 @@ import json
 import random
 
 
-# 可視化用クラスラベル
 oid_labels = [
   'Person',
   'Bicycle',
@@ -21,11 +20,9 @@ oid_labels = [
   'Drink'
 ]
 
-# 可視化用COLOR
 COLORS = [[0.000, 0.447, 0.741], [0.850, 0.325, 0.098], [0.929, 0.694, 0.125],
           [0.494, 0.184, 0.556], [0.466, 0.674, 0.188], [0.301, 0.745, 0.933]]
 
-# 標準的なPyTorchのmean-std入力画像の正規化
 transform = T.Compose([
     T.Resize(800),
     T.ToTensor(),
@@ -33,36 +30,25 @@ transform = T.Compose([
 ])
 
 def box_cxcywh_to_xyxy(x):
-    """
-    (center_x, center_y, width, height)から(xmin, ymin, xmax, ymax)に座標変換
-    """
-    # unbind(1)でTensor次元を削除
-    # (center_x, center_y, width, height)*N → (center_x*N, center_y*N, width*N, height*N)
+    # (center_x, center_y, width, height)*N -> (center_x*N, center_y*N, width*N, height*N)
     x_c, y_c, w, h = x.unbind(1)
     b = [(x_c - 0.5 * w), (y_c - 0.5 * h), (x_c + 0.5 * w), (y_c + 0.5 * h)]
     # (center_x, center_y, width, height)*N の形に戻す
     return torch.stack(b, dim=1)
 
 def rescale_bboxes(out_bbox, size):
-    """
-    バウンディングボックスのリスケール
-    """
     img_w, img_h = size
     b = box_cxcywh_to_xyxy(out_bbox)
-    # バウンディングボックスの[0～1]から元画像の大きさにリスケール
     b = b * torch.tensor([img_w, img_h, img_w, img_h], dtype=torch.float32)
     return b
 
 def filter_bboxes_from_outputs(outputs, im, threshold=0.75):
-  # 閾値以上の信頼度を持つ予測値のみを保持
   probas = outputs['pred_logits'].softmax(-1)[0, :, :-1]
   keep = probas.max(-1).values > threshold
   probas_to_keep = probas[keep]
-  # [0, 1]のボックスを画像のスケールに変換
   bboxes_scaled = rescale_bboxes(outputs['pred_boxes'][0, keep], im.size)
   return probas_to_keep, bboxes_scaled
 
-# 結果の表示
 def plot_finetuned_results(pil_img, image_name, threshold, prob=None, boxes=None, labels=None):
   plt.figure(figsize=(16, 10))
   plt.imshow(pil_img)
@@ -80,11 +66,10 @@ def plot_finetuned_results(pil_img, image_name, threshold, prob=None, boxes=None
   plt.axis('off')
   plt.savefig(f'./outputs/{image_name}_{device}_{epoch}_{threshold}.png')
 
-# 物体検出
+# object detection
 def run_worflow(my_image, image_name, my_model, labels, device, epoch, topology, threshold=0.75):
-  # mean-std入力画像の正規化(バッチサイズ : 1)
+  # normalization of mean-std input images (batch size : 1)
   img = transform(my_image).unsqueeze(0)
-  # モデルに反映
   outputs = my_model(img)
 
   probas_to_keep, bboxes_scaled = filter_bboxes_from_outputs(outputs, my_image, threshold=threshold)
@@ -95,7 +80,7 @@ if __name__ == "__main__":
   in_features = model.class_embed.in_features
   model.class_embed = nn.Linear(in_features=in_features, out_features=11)
 
-  # please edit these configuration
+  # please edit these configurations
   device = 5
   epoch = '0049'
   threshold = 0.75
@@ -104,5 +89,5 @@ if __name__ == "__main__":
   checkpoint = torch.load(f'./outputs/node{device}/checkpoint{epoch}.pth', map_location='cpu')
   model.load_state_dict(checkpoint['model'], strict=False)
 
-  im = Image.open(f'./data/custom/val2017/{image_name}')
+  im = Image.open(f'../data/custom/val2017/{image_name}')
   run_worflow(im, image_name, model, oid_labels, device, epoch, threshold)
